requirements.txt
ansicon==1.89.0
anyio==3.7.1
arrow==1.2.3
asgiref==3.7.2
attrs==23.1.0
black==23.3.0
blessed==1.20.0
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.6
Django==4.0.4
django-cors-headers==4.1.0
django-debug-toolbar==3.4.0
django-picklefield==3.1
django-q==1.3.9
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.22.1
flake8==5.0.4
flake8-annotations==2.9.1
flake8-quotes==3.3.1
flake8-variables-names==0.0.5
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
inflection==0.5.1
jinxed==1.2.0
jsonschema==4.17.3
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
pep8-naming==0.13.2
pexpect==4.8.0
Pillow==10.0.0
platformdirs==3.8.0
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.7.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.4
pytz==2023.3
PyYAML==6.0
redis==3.5.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
Strip==0.0.1.dev14
stripe==5.4.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
wcwidth==0.2.6
